Update prices and amounts to use uint instead of floats
Former-commit-id: d88bb3bdbd901ada0031a091ee4503be3ecac097

+++ trading_engine/book_entry.go
@@ -2,8 +2,8 @@ package trading_engine
 
 // BookEntry keeps minimal information about the order in the OrderBook
 type BookEntry struct {
-	Price  float64
-	Amount float64
+	Price  uint64
+	Amount uint64
 	Order  Order
 }
